# packages/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TODOMaster API
  description: |
    TODOMaster是一个基于OKR（Objectives and Key Results）方法论的任务管理系统。
    本API提供了完整的用户管理、目标设置、关键结果跟踪和任务管理功能。
  version: 1.0.0
  contact:
    name: TODOMaster Team
    email: support@todomaster.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 开发服务器
  - url: https://api.todomaster.com/v1
    description: 生产服务器
...
security:
  - bearerAuth: []
paths:
...
  /key-results:
    get:
      tags: [key-results]
      summary: 获取关键结果列表
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: objectiveId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, archived]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 关键结果列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyResultsListResponse'

    post:
      tags: [key-results]
      summary: 创建关键结果
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKeyResultRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyResultDetailResponse'
  /key-results/{id}:
    get:
      tags: [key-results]
      summary: 获取关键结果详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 关键结果详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyResultDetailResponse'

    put:
      tags: [key-results]
      summary: 更新关键结果
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateKeyResultRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyResultDetailResponse'

    delete:
      tags: [key-results]
      summary: 删除关键结果
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    KeyResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [number, percentage, boolean]
        currentValue:
          type: number
        targetValue:
          type: number
        unit:
          type: string
        progress:
          type: number
          minimum: 0
          maximum: 100
        status:
          type: string
          enum: [active, completed, archived]
        objectiveId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer

    # 请求模型
...
    CreateKeyResultRequest:
      type: object
      required:
        - title
        - type
        - targetValue
        - objectiveId
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [number, percentage, boolean]
        targetValue:
          type: number
        currentValue:
          type: number
          default: 0
        unit:
          type: string
        objectiveId:
          type: string
          format: uuid
    UpdateKeyResultRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        targetValue:
          type: number
        currentValue:
          type: number
        unit:
          type: string
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
    KeyResultsListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                keyResults:
                  type: array
                  items:
                    $ref: '#/components/schemas/KeyResult'
                pagination:
                  $ref: '#/components/schemas/Pagination'
    KeyResultDetailResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                keyResult:
                  $ref: '#/components/schemas/KeyResult'
...